git-svn-id: https://robotics.mvla.net/svn/frc971/2013/trunk/src@4140 f308d9b7-e957-4cde-b6ac-9a88185e7312

+++ 971CV/src/org/frc971/ImageWithTimestamp.java
@@ -4,5 +4,5 @@ import com.googlecode.javacv.cpp.opencv_core.IplImage;
 
 public class ImageWithTimestamp {
 	IplImage image;
-	float timestamp;
+	double timestamp;
 }
